+++ Client/ClientCode.py
@@ -6,7 +6,7 @@ from PyQt5 import QtCore
 from PyQt5.QtCore import QPropertyAnimation, Qt, QTimer
 from PyQt5.QtWidgets import QMainWindow, QApplication, QListView
 
-from Client.Bubble.LabelBubble import MessageDelegate, MessageModel, USER_ME, USER_THEM, USER_ADMIN
+from Client.Bubble.LabelBubble import MessageDelegate, MessageModel, USER_ME, USER_THEM, USER_ADMIN,BUBBLE_COLORS
 from Client.Username.Choose_Draggable import Draggable
 from Client_UI import Ui_MainWindow
 
@@ -16,8 +16,13 @@ from time import time
 
 HOST = '127.0.0.1'
 PORT = 9090
-#Server Messages
+# Server Messages
 s_messages = ('connected to the server!', 'Disconnected from the server!')
+import random
+
+def rand_color():
+    r = lambda: random.randint(0,255)
+    return  '#%02X%02X%02X' % (r(),r(),r())
 
 
 class ClientCode(Ui_MainWindow, QMainWindow):
@@ -61,10 +66,9 @@ class ClientCode(Ui_MainWindow, QMainWindow):
         message = f"{self.nickname}:{self.textEdit.toPlainText()}\n"
         self.sock.send(message.encode('UTF-8'))
         r_nickname = message.split(':')[0]
-        r_message = message.split(':')[-1]
         # Check if username is in message by splitting up
         if self.nickname == r_nickname:
-            self.model.add_message(USER_ME, message.split(':')[-1], time(),message.split(':')[0])
+            self.model.add_message(USER_ME, message.split(':')[-1], time(),message.split(':')[0],"#90caf9")
         self.textEdit.clear()
 
     def receive(self):
@@ -78,12 +82,12 @@ class ClientCode(Ui_MainWindow, QMainWindow):
                 if message == 'NICK':
                     self.sock.send(self.nickname.encode('UTF-8'))
                 elif any(check in message for check in s_messages):
-                    self.model.add_message(USER_ADMIN, r_message, time(),r_nickname)
+                    self.model.add_message(USER_ADMIN, r_message, time(),r_nickname,"#FFFFFF")
                 else:
                     if self.gui_done:
                         self.textBrowser.insertPlainText(message + "\n")
                         if self.nickname != r_nickname:
-                            self.model.add_message(USER_THEM,r_message, time(), r_nickname)
+                            self.model.add_message(USER_THEM,r_message, time(), r_nickname,rand_color())
             except ConnectionAbortedError:
                 break
             except: